+++ pom.xml
@@ -4,7 +4,7 @@
 	<parent>
 		<groupId>org.matsim</groupId>
 		<artifactId>matsim-all</artifactId>
-		<version>15.0-PR2457</version>
+		<version>16.0-PR2530</version>
 	</parent>
 
 	<modelVersion>4.0.0</modelVersion>
