+++ src/test/java/org/matsim/run/ChessboardParkingTest.java
@@ -11,7 +11,6 @@ import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
-import org.matsim.api.core.v01.network.Node;
 import org.matsim.api.core.v01.population.*;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
@@ -20,10 +19,8 @@ import org.matsim.core.config.groups.StrategyConfigGroup;
 import org.matsim.core.config.groups.VspExperimentalConfigGroup;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.controler.Controler;
-import org.matsim.core.network.NetworkUtils;
 import org.matsim.core.population.PopulationUtils;
 import org.matsim.core.population.algorithms.PlanAlgorithm;
-import org.matsim.core.population.routes.NetworkRoute;
 import org.matsim.core.replanning.PlanStrategy;
 import org.matsim.core.replanning.PlanStrategyImpl;
 import org.matsim.core.replanning.modules.AbstractMultithreadedModule;
@@ -33,22 +30,14 @@ import org.matsim.core.scenario.MutableScenario;
 import org.matsim.core.scenario.ScenarioUtils;
 import org.matsim.core.utils.io.IOUtils;
 import org.matsim.core.utils.timing.TimeInterpretation;
-import org.matsim.core.utils.timing.TimeTracker;
 import org.matsim.examples.ExamplesUtils;
 import org.matsim.facilities.ActivityFacilities;
-import org.matsim.facilities.FacilitiesUtils;
-import org.matsim.facilities.Facility;
 import org.matsim.run.prepare.LeipzigUtils;
 import org.matsim.testcases.MatsimTestUtils;
-import org.matsim.vehicles.Vehicle;
 import playground.vsp.openberlinscenario.cemdap.output.ActivityTypes;
 
 import javax.inject.Provider;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.matsim.run.prepare.LeipzigUtils.parkingIsRestricted;
 
 /**
  * abc
@@ -108,169 +97,6 @@ public class ChessboardParkingTest {
 		controler.run();
 	}
 
-	static final class LeipzigPlanRouter implements PlanAlgorithm {
-		private final TripRouter tripRouter;
-		private final ActivityFacilities facilities;
-		private final TimeInterpretation timeInterpretation;
-		private final Network fullModalNetwork;
-		private final Network reducedNetwork;
-		private final MultimodalLinkChooser linkChooser;
-		private final Scenario scenario;
-		LeipzigPlanRouter( final TripRouter tripRouter, final ActivityFacilities facilities, final TimeInterpretation timeInterpretation,
-					  SingleModeNetworksCache singleModeNetworksCache, Scenario scenario, MultimodalLinkChooser linkChooser ) {
-			this.tripRouter = tripRouter;
-			this.facilities = facilities;
-			this.timeInterpretation = timeInterpretation;
-			this.scenario = scenario;
-			this.fullModalNetwork = singleModeNetworksCache.getSingleModeNetworksCache().get( TransportMode.car );
-
-			// yyyy one should look at the networks cache and see how the following is done.  And maybe even register it there.
-			this.reducedNetwork = NetworkUtils.createNetwork( scenario.getConfig().network() );
-			this.linkChooser = linkChooser;
-			for( Node node : this.fullModalNetwork.getNodes().values() ){
-				reducedNetwork.addNode( node );
-			}
-			for( Link link : this.fullModalNetwork.getLinks().values() ){
-				if ( !LeipzigUtils.parkingIsRestricted( link ) ) {
-					reducedNetwork.addLink( link );
-				}
-			}
-			log.warn("returning LeipzigPlanRouter");
-		}
-		enum ParkingType { normal, restricted, shopping }
-		@Override public void run(final Plan plan) {
-			final List<TripStructureUtils.Trip> trips = TripStructureUtils.getTrips( plan );
-			TimeTracker timeTracker = new TimeTracker(timeInterpretation);
-
-			log.warn( "=== old plan: ===" );
-			for( PlanElement tripElement : plan.getPlanElements() ){
-				log.warn( tripElement );
-			}
-			log.warn( "======" );
-
-			for ( TripStructureUtils.Trip oldTrip : trips) {
-				final String routingMode = TripStructureUtils.identifyMainMode( oldTrip.getTripElements() );
-				timeTracker.addActivity(oldTrip.getOriginActivity());
-
-				final Facility fromFacility = FacilitiesUtils.toFacility( oldTrip.getOriginActivity(), facilities );
-				final Facility toFacility = FacilitiesUtils.toFacility( oldTrip.getDestinationActivity(), facilities );
-
-				// At this point, I only want to deal with residential parking.  Shopping comes later (and is simpler).
-
-				ParkingType parkingTypeAtOrigin = getParkingType( fullModalNetwork, oldTrip.getOriginActivity() );
-				ParkingType parkingTypeAtDestination = getParkingType( fullModalNetwork, oldTrip.getDestinationActivity() );
-
-				if ( parkingTypeAtOrigin==ParkingType.normal  && parkingTypeAtDestination==ParkingType.normal ){
-					// standard case:
-
-					final List<? extends PlanElement> newTripElements = tripRouter.calcRoute( routingMode, fromFacility, toFacility,
-							timeTracker.getTime().seconds(), plan.getPerson(), oldTrip.getTripAttributes() );
-					putVehicleFromOldTripIntoNewTripIfMeaningful( oldTrip, newTripElements );
-					TripRouter.insertTrip( plan, oldTrip.getOriginActivity(), newTripElements, oldTrip.getDestinationActivity() );
-					timeTracker.addElements( newTripElements );
-				} else if ( parkingTypeAtOrigin==ParkingType.restricted && parkingTypeAtDestination==ParkingType.normal ) {
-					// restricted parking at origin:
-
-					// first find parking:
-//					final Link parkingLink = NetworkUtils.getNearestLink( reducedNetwork, oldTrip.getOriginActivity().getCoord() );
-
-					final Link parkingLink = linkChooser.decideOnLink( fromFacility, reducedNetwork );
-
-					final Activity parkingActivity = scenario.getPopulation().getFactory().createInteractionActivityFromLinkId(
-							TripStructureUtils.createStageActivityType( "parking" ), parkingLink.getId() );
-					final Facility parkingFacility = FacilitiesUtils.toFacility( parkingActivity, facilities );
-
-					List<PlanElement> newTripElements = new ArrayList<>();
-
-					// trip from origin to parking:
-					final List<? extends PlanElement> walkTripElements = tripRouter.calcRoute( TransportMode.walk, fromFacility, parkingFacility,
-							timeTracker.getTime().seconds(), plan.getPerson(), oldTrip.getTripAttributes() );
-					newTripElements.addAll( walkTripElements );
-
-					// parking interaction:
-					newTripElements.add( parkingActivity );
-
-					// trip from parking to final destination:
-					final List<? extends PlanElement> carTripElements = tripRouter.calcRoute( routingMode, parkingFacility, toFacility,
-							timeTracker.getTime().seconds(), plan.getPerson(), oldTrip.getTripAttributes() );
-					newTripElements.addAll( carTripElements );
-
-					putVehicleFromOldTripIntoNewTripIfMeaningful( oldTrip, newTripElements );
-					TripRouter.insertTrip( plan, parkingActivity, newTripElements, oldTrip.getDestinationActivity() );
-					timeTracker.addElements( newTripElements );
-
-
-				} else  {
-					throw new RuntimeException();
-					// to be implemented!
-				}
-
-			}
-			log.warn( "=== new plan: ===" );
-			for( PlanElement tripElement : plan.getPlanElements() ){
-				log.warn( tripElement );
-			}
-			log.warn( "======" );
-
-		}
-		private static ParkingType getParkingType( Network fullModalNetwork, Activity originActivity ){
-			ParkingType parkingTypeAtOrigin = ParkingType.normal;
-
-			// check if non-home activity (since otherwise we assume that there is no parking restriction):
-			if ( !originActivity.getType().equals( ActivityTypes.HOME ) ) {
-
-				// if non-home activity, check if in restricted zone:
-				Link link = fullModalNetwork.getLinks().get( originActivity.getLinkId() );
-				if ( parkingIsRestricted(link) ) {
-					parkingTypeAtOrigin = ParkingType.restricted;
-				}
-
-			}
-			return parkingTypeAtOrigin;
-		}
-
-		/**
-		 * If the old trip had vehicles set in its network routes, and it used a single vehicle,
-		 * and if the new trip does not come with vehicles set in its network routes,
-		 * then put the vehicle of the old trip into the network routes of the new trip.
-		 * @param oldTrip The old trip
-		 * @param newTrip The new trip
-		 * @deprecated -- use from PlanRouter
-		 */
-		private static void putVehicleFromOldTripIntoNewTripIfMeaningful( TripStructureUtils.Trip oldTrip, List<? extends PlanElement> newTrip ) {
-			Id<Vehicle> oldVehicleId = getUniqueVehicleId(oldTrip );
-			if (oldVehicleId != null) {
-				for (Leg leg : TripStructureUtils.getLegs(newTrip)) {
-					if (leg.getRoute() instanceof NetworkRoute ) {
-						if (((NetworkRoute) leg.getRoute()).getVehicleId() == null) {
-							((NetworkRoute) leg.getRoute()).setVehicleId(oldVehicleId);
-						}
-					}
-				}
-			}
-		}
-
-		/**
-		 * @param trip
-		 * @return
-		 * 		 * @deprecated -- use from PlanRouter
-		 */
-		private static Id<Vehicle> getUniqueVehicleId( TripStructureUtils.Trip trip ) {
-			Id<Vehicle> vehicleId = null;
-			for (Leg leg : trip.getLegsOnly()) {
-				if (leg.getRoute() instanceof NetworkRoute) {
-					if (vehicleId != null && (!vehicleId.equals(((NetworkRoute) leg.getRoute()).getVehicleId()))) {
-						return null; // The trip uses several vehicles.
-					}
-					vehicleId = ((NetworkRoute) leg.getRoute()).getVehicleId();
-				}
-			}
-			return vehicleId;
-		}
-
-
-	}
-
 	static final class LeipzigRoutingStrategyProvider implements Provider<PlanStrategy> {
 		// is a provider in matsim core.  maybe try without.  kai, apr'23
 		@Inject private GlobalConfigGroup globalConfigGroup;
@@ -284,7 +110,7 @@ public class ChessboardParkingTest {
 			PlanStrategyImpl.Builder builder = new PlanStrategyImpl.Builder( new RandomPlanSelector<>()) ;
 			builder.addStrategyModule( new AbstractMultithreadedModule( globalConfigGroup ){
 				@Override public final PlanAlgorithm getPlanAlgoInstance() {
-					return new LeipzigPlanRouter( tripRouterProvider.get(), facilities, timeInterpretation, singleModeNetworksCache, scenario, linkChooser );
+					return new LeipzigRouterPlanAlgorithm( tripRouterProvider.get(), facilities, timeInterpretation, singleModeNetworksCache, scenario, linkChooser );
 				}
 			} );
 			return builder.build() ;
